+++ lua/mini/tabline.lua
@@ -101,9 +101,26 @@ end
 ---
 --- Default values:
 ---@eval return MiniDoc.afterlines_to_code(MiniDoc.current.eval_section)
+---@text # Format ~
+---
+--- `config.format` is a callable that takes buffer identifier and pre-computed
+--- label as arguments and returns a string with formatted label.
+--- This function will be called for all displayable in tabline buffers.
+--- Default: |MiniTabline.default_format()|.
+---
+--- Example of adding "+" suffix for modified buffers: >
+---
+---   function(buf_id, label)
+---     local suffix = vim.bo[buf_id].modified and '+ ' or ''
+---     return MiniTabline.default_format(buf_id, label) .. suffix
+---   end
 MiniTabline.config = {
   -- Whether to show file icons (requires 'nvim-tree/nvim-web-devicons')
   show_icons = true,
+
+  -- Function which formats the tab label
+  -- By default surrounds with space and possibly prepends with icon
+  format = nil,
 
   -- Whether to set Vim's settings for tabline (make it always shown and
   -- allow hidden buffers)
@@ -126,6 +143,22 @@ MiniTabline.make_tabline_string = function()
   H.fit_width()
 
   return H.concat_tabs()
+end
+
+--- Default tab format
+---
+--- Used by default as `config.format`.
+--- Prepends label with padded icon (if `show_icon` from |MiniTabline.config|
+--- is `true`) and surrounds label with single space.
+--- Note: it is meant to be used only as part of `format` in |MiniTabline.config|.
+---
+---@param buf_id number Buffer identifier.
+---@param label string Pre-computed label.
+---
+---@return string Formatted label.
+MiniTabline.default_format = function(buf_id, label)
+  if H.get_icon == nil then return string.format(' %s ', label) end
+  return string.format(' %s %s ', H.get_icon(label), label)
 end
 
 -- Helper data ================================================================
@@ -160,6 +193,7 @@ H.setup_config = function(config)
 
   vim.validate({
     show_icons = { config.show_icons, 'boolean' },
+    format = { config.format, 'function', true },
     set_vim_settings = { config.set_vim_settings, 'boolean' },
     tabpage_section = { config.tabpage_section, 'string' },
   })
@@ -352,25 +386,26 @@ H.finalize_labels = function()
     nonunique_tab_ids = H.get_nonunique_tab_ids()
   end
 
-  -- Postprocess: add file icons and padding
-  local has_devicons, devicons
-  local show_icons = H.get_config().show_icons
+  -- Format labels
+  local config = H.get_config()
 
-  -- Have this `require()` here to not depend on plugin initialization order
-  if show_icons then
-    has_devicons, devicons = pcall(require, 'nvim-web-devicons')
+  -- - Ensure cached `get_icon` for `default_format` (for better performance)
+  if not config.show_icons then H.get_icon = nil end
+  if config.show_icons and H.get_icon == nil then
+    -- Have this `require()` here to not depend on plugin initialization order
+    local has_devicons, devicons = pcall(require, 'nvim-web-devicons')
+    if has_devicons then H.get_icon = function(name) return devicons.get_icon(name, nil, { default = true }) end end
   end
 
+  -- - Apply formatting
+  local format = config.format or MiniTabline.default_format
   for _, tab in pairs(H.tabs) do
-    if show_icons and has_devicons then
-      local extension = vim.fn.fnamemodify(tab.label, ':e')
-      local icon = devicons.get_icon(tab.label, extension, { default = true })
-      tab.label = string.format(' %s %s ', icon, tab.label)
-    else
-      tab.label = string.format(' %s ', tab.label)
-    end
+    tab.label = format(tab.buf_id, tab.label)
   end
 end
+
+-- Helper for `default_format`, cached for performance
+H.get_icon = nil
 
 ---@return table Array of `H.tabs` ids which have non-unique labels.
 ---@private