+++ tests/dir-tabline/lua/nvim-web-devicons.lua
@@ -3,8 +3,8 @@ return {
     _G.devicons_args = { filename = filename, extension = extension, options = options }
 
     if filename == 'LICENSE' then return '', 'DevIconLicense' end
-    if extension == 'lua' then return '', 'DevIconLua' end
-    if extension == 'txt' then return '', 'DevIconTxt' end
+    if vim.endswith(filename, 'lua') then return '', 'DevIconLua' end
+    if vim.endswith(filename, 'txt') then return '', 'DevIconTxt' end
     if (options or {}).default then return '', 'DevIconDefault' end
   end,
 }